update offline info

+++ src/services/file-system-api.js
@@ -1,17 +1,23 @@
 import PhoneStorage from "./phone-storage";
 import * as FileSystem from "expo-file-system";
+import { Alert } from "react-native";
+import NetInfo from "@react-native-community/netinfo";
 const cacheDirectory = FileSystem.cacheDirectory + "sv1712324/";
 const appDirectory = FileSystem.documentDirectory + "sv1712324/";
-const getFileExt = (filePath) => filePath.includes("lh3.googleusercontent.com") ? "png" : filePath.substring(filePath.lastIndexOf(".") + 1);
+const getFileExt = (filePath) =>
+  filePath.includes("lh3.googleusercontent.com")
+    ? "png"
+    : filePath.substring(filePath.lastIndexOf(".") + 1);
 const courseBaseDir = () => appDirectory + `Course/`;
 const courseDir = (course_id) => appDirectory + `Course/${course_id}/`;
-const courseVideoDir = (course_id, ext) => courseDir(course_id) + `video.${ext}`;
+const courseVideoDir = (course_id, ext) =>
+  courseDir(course_id) + `video.${ext}`;
 const courseImageDir = (course_id, image_ext) =>
   courseDir(course_id) + `image.${image_ext}`;
 const lessonDir = (course_id, lesson_id, ext) =>
   courseDir(course_id) + `lesson_${lesson_id}.${ext}`;
-  const instructorDir = (instructor_id) =>
-    appDirectory + `instructor_${instructor_id}/`;
+const instructorDir = (instructor_id) =>
+  appDirectory + `instructor_${instructor_id}/`;
 const instructorImageDir = (instructor_id, fileExt) =>
   instructorDir(instructor_id) + `image.${fileExt}`;
 const createDir = async (dirPath) => {
@@ -24,8 +30,12 @@ const downloadCourse = async (courseDownloadInfo, course_id) => {
   try {
     await createDir(courseDir(course_id));
 
-    console.log("Downloading files...");
-    await Promise.all(
+    NetInfo.addEventListener((state) => {});
+    const state = await NetInfo.fetch();
+    if (!state.isInternetReachable) {
+      return null;
+    }
+    return await Promise.all(
       courseDownloadInfo.map((data) => {
         const filePath =
           data.type == "course"
@@ -36,16 +46,24 @@ const downloadCourse = async (courseDownloadInfo, course_id) => {
         FileSystem.getInfoAsync(filePath).then((fileInfo) => {
           if (!fileInfo.exists) {
             FileSystem.downloadAsync(data.path, filePath);
+          } else {
+            console.log("exists");
           }
         });
       })
     );
   } catch (e) {
     console.error("Couldn't download course files:", e);
+    return null;
   }
 };
 const downloadCourseImage = async (course_id, imagePath) => {
   try {
+    NetInfo.addEventListener((state) => {});
+    const state = await NetInfo.fetch();
+    if (!state.isInternetReachable) {
+      return null;
+    }
     await createDir(courseDir(course_id));
 
     const imageExt = getFileExt(imagePath);
@@ -61,26 +79,42 @@ const downloadCourseImage = async (course_id, imagePath) => {
 const downloadInstructorImage = async (instructor_id, imagePath) => {
   try {
     await createDir(instructorDir(instructor_id));
+    NetInfo.addEventListener((state) => {});
+    const state = await NetInfo.fetch();
+    if (!state.isInternetReachable) {
+      return null;
+    }
     const imageExt = getFileExt(imagePath);
     const fileUri = instructorImageDir(instructor_id, imageExt);
     const fileInfo = await FileSystem.getInfoAsync(fileUri);
-    if (fileInfo.exists) {
-      await FileSystem.deleteAsync(fileUri);
-    }
+    try {
+      if (fileInfo.exists) {
+        await FileSystem.deleteAsync(fileUri);
+      }
+    } catch (err) {}
     await FileSystem.downloadAsync(imagePath, fileUri);
   } catch (e) {
     console.error("Couldn't download image file:", e);
   }
 };
 const deleteAllCourse = async () => {
-  await FileSystem.deleteAsync(courseBaseDir());
+  return await FileSystem.deleteAsync(courseBaseDir());
 };
 const deleteCourse = async (course_id) => {
-  await FileSystem.deleteAsync(courseDir(course_id));
-}
+  return await FileSystem.deleteAsync(courseDir(course_id));
+};
 const downloadImageCouseList = async (payloadStr) => {
   const payload = JSON.parse(payloadStr);
-  Promise.all(
+  NetInfo.addEventListener((state) => {});
+  try {
+    const state = await NetInfo.fetch();
+    if (!state.isInternetReachable) {
+      return null;
+    }
+  } catch (err) {
+    return null;
+  }
+  return Promise.all(
     payload.map((course) =>
       course.imageUrl
         ? FileSystemApi.downloadCourseImage(course.id, course.imageUrl)
@@ -89,7 +123,7 @@ const downloadImageCouseList = async (payloadStr) => {
         : null
     )
   );
-}
+};
 const getCourseVideo = async (course_id, filePath) => {
   const fileUri = courseVideoDir(course_id, getFileExt(filePath));
   const fileInfo = await FileSystem.getInfoAsync(fileUri);
@@ -97,7 +131,7 @@ const getCourseVideo = async (course_id, filePath) => {
     return fileUri;
   }
   return null;
-}
+};
 const getCourseImage = async (course_id, filePath) => {
   const fileUri = courseImageDir(course_id, getFileExt(filePath));
   const fileInfo = await FileSystem.getInfoAsync(fileUri);
